requirements.txt
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.3
cssselect2==0.8.0
Deprecated==1.2.18
deprecation==2.1.0
ecdsa==0.19.1
etelemetry==0.3.1
fastapi==0.115.12
filelock==3.18.0
firebase-admin==6.8.0
fonttools==4.58.0
frozenlist==1.6.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
jwt==1.3.1
limits==5.1.0
linkify-it-py==2.0.3
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20250506
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==4.0.0
pydyf==0.11.0
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
realtime==2.4.3
reportlab==4.4.0
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
secure==0.2.1
simplejson==3.20.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tinycss2==1.4.0
tinyhtml5==2.0.0
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
weasyprint==65.1
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
yarl==1.20.0
zopfli==0.2.3.post1
